+++ Solutions/Day9.cxx
@@ -16,12 +16,12 @@ std::string Day9::solve(Part part) const {
   for (int i = 0; i < m_input.size(); i++) {
     std::vector<int> partition;
     if (i % 2 == 0) {
-      for (int j = 0; j < m_input.at(i); ++j) {
+      for (size_t j = 0; j < m_input.at(i); ++j) {
         layout.push_back(i / 2);
         partition.push_back(i / 2);
       }
     } else {
-      for (int j = 0; j < m_input.at(i); ++j) {
+      for (size_t j = 0; j < m_input.at(i); ++j) {
         layout.push_back(-1);
         partition.push_back(-1);
       }
@@ -33,14 +33,14 @@ std::string Day9::solve(Part part) const {
 
   if (part == PART_1) {
     auto start = layout.begin();
-    for (int i = layout.size() - 1; i >= 0; --i) {
+    for (size_t i = layout.size() - 1; i >= 0; --i) {
       start = std::find(start, layout.end(), -1);
-      if (start != layout.end() && i > (start - layout.begin())) {
+      if (start != layout.end() && i > static_cast<size_t>(start - layout.begin())) {
         std::swap(layout.at(i), layout.at(start - layout.begin()));
       }
     }
 
-    for (int i = 0; i < layout.size(); ++i) {
+    for (size_t i = 0; i < layout.size(); ++i) {
       if (layout.at(i) > 0) {
         total += i * layout.at(i);
       }
@@ -49,7 +49,7 @@ std::string Day9::solve(Part part) const {
 
   if (part == PART_2) {
     auto start = partitions.begin();
-    for (int i = partitions.size() - 1; i >= 0; --i) {
+    for (size_t i = partitions.size() - 1; i >= 0; --i) {
       std::vector<int>& file = partitions.at(i);
       if (file.empty() || file.at(0) == -1) {
         continue;
@@ -58,13 +58,13 @@ std::string Day9::solve(Part part) const {
         return partition.size() >= file.size() && partition.at(0) == -1;
       };
       auto it = std::find_if(partitions.begin(), partitions.end(), partitionMatcher);
-      if (it != partitions.end() && i > (it - partitions.begin())) {
+      if (it != partitions.end() && i > static_cast<size_t>(it - partitions.begin())) {
         std::vector<int>& space = partitions.at(it - partitions.begin());
         if (file.size() == space.size()) {
           std::swap(file, space);
         } else {
           std::vector<int> difference;
-          for (int j = file.size(); j < space.size(); ++j) {
+          for (size_t j = file.size(); j < space.size(); ++j) {
             difference.push_back(-1);
           }
           space.resize(file.size());
